+++ x/perp/v2/keeper/amm_test.go
@@ -160,14 +160,17 @@ func TestShiftPegMultiplier_Fail(t *testing.T) {
 
 	account := testutil.AccAddress()
 
-	err := app.PerpKeeperV2.Sudo().CreateMarket(
+	adminUser, err := sdk.AccAddressFromBech32(testutil.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
+	err = app.PerpKeeperV2.Sudo().CreateMarket(
 		ctx,
 		keeper.ArgsCreateMarket{
 			Pair:            pair,
 			PriceMultiplier: sdk.NewDec(2),
 			SqrtDepth:       sdk.NewDec(1_000_000),
-			EnableMarket:    true,
 		},
+		adminUser,
 	)
 	app.PerpKeeperV2.ReserveSnapshots.Insert(
 		ctx,
@@ -230,14 +233,17 @@ func TestShiftSwapInvariant_Fail(t *testing.T) {
 	app, ctx := testapp.NewNibiruTestAppAndContext()
 	account := testutil.AccAddress()
 
-	err := app.PerpKeeperV2.Sudo().CreateMarket(
+	adminUser, err := sdk.AccAddressFromBech32(testutil.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
+	err = app.PerpKeeperV2.Sudo().CreateMarket(
 		ctx,
 		keeper.ArgsCreateMarket{
 			Pair:            pair,
 			PriceMultiplier: sdk.NewDec(2),
 			SqrtDepth:       sdk.NewDec(1_000),
-			EnableMarket:    true,
 		},
+		adminUser,
 	)
 	app.PerpKeeperV2.ReserveSnapshots.Insert(
 		ctx,
@@ -437,14 +443,17 @@ func TestKeeper_GetMarketByPairAndVersion(t *testing.T) {
 
 	pair := asset.Registry.Pair(denoms.BTC, denoms.NUSD)
 
-	err := app.PerpKeeperV2.Sudo().CreateMarket(
+	adminUser, err := sdk.AccAddressFromBech32(testutil.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
+	err = app.PerpKeeperV2.Sudo().CreateMarket(
 		ctx,
 		keeper.ArgsCreateMarket{
 			Pair:            pair,
 			PriceMultiplier: sdk.NewDec(2),
 			SqrtDepth:       sdk.NewDec(1_000_000),
-			EnableMarket:    true,
 		},
+		adminUser,
 	)
 	require.NoError(t, err)
 
@@ -462,14 +471,17 @@ func TestKeeper_GetAMMByPairAndVersion(t *testing.T) {
 
 	pair := asset.Registry.Pair(denoms.BTC, denoms.NUSD)
 
-	err := app.PerpKeeperV2.Sudo().CreateMarket(
+	adminUser, err := sdk.AccAddressFromBech32(testutil.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
+	err = app.PerpKeeperV2.Sudo().CreateMarket(
 		ctx,
 		keeper.ArgsCreateMarket{
 			Pair:            pair,
 			PriceMultiplier: sdk.NewDec(2),
 			SqrtDepth:       sdk.NewDec(1_000_000),
-			EnableMarket:    true,
 		},
+		adminUser,
 	)
 	require.NoError(t, err)