+++ x/perp/v2/module/abci_test.go
@@ -25,6 +25,9 @@ func TestSnapshotUpdates(t *testing.T) {
 	initialMarket := *mock.TestMarket()
 	initialAmm := *mock.TestAMMDefault()
 
+	adminUser, err := sdk.AccAddressFromBech32(testutilevents.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
 	runBlock := func(duration time.Duration) {
 		perp.EndBlocker(ctx, app.PerpKeeperV2)
 		ctx = ctx.
@@ -41,6 +44,7 @@ func TestSnapshotUpdates(t *testing.T) {
 			SqrtDepth:       initialAmm.SqrtDepth,
 			Market:          &initialMarket,
 		},
+		adminUser,
 	))
 
 	expectedSnapshot := types.ReserveSnapshot{
@@ -108,6 +112,9 @@ func TestEndBlocker(t *testing.T) {
 	initialMarket := *mock.TestMarket()
 	initialAmm := *mock.TestAMMDefault()
 
+	adminUser, err := sdk.AccAddressFromBech32(testutilevents.ADDR_SUDO_ROOT)
+	require.NoError(t, err)
+
 	runBlock := func(duration time.Duration) {
 		perp.EndBlocker(ctx, app.PerpKeeperV2)
 		ctx = ctx.
@@ -126,6 +133,7 @@ func TestEndBlocker(t *testing.T) {
 			SqrtDepth:       initialAmm.SqrtDepth,
 			Market:          &initialMarket,
 		},
+		adminUser,
 	))
 
 	t.Log("run one block of 5 seconds")