+++ x/sudo/cli/cli_test.go
@@ -159,7 +159,7 @@ func (s *IntegrationSuite) TestCmdEditSudoers() {
 		contracts = append(contracts, addr.String())
 	}
 
-	var sender sdk.AccAddress = s.root.addr
+	var sender = s.root.addr
 
 	pbMsg := types.MsgEditSudoers{
 		Action:    "add_contracts",